refactor base components to new directory
