+++ ohmg/frontend/svelte_components/src/components/cards/LayerCard.svelte
@@ -4,11 +4,11 @@
   import DownloadSimple from 'phosphor-svelte/lib/DownloadSimple';
   import MapPin from 'phosphor-svelte/lib/MapPin';
 
-  import { getModal } from '../modals/BaseModal.svelte';
+  import { getModal } from '../base/Modal.svelte';
 
   import Link from '../common/Link.svelte';
 
-  import BaseCard from './BaseCard.svelte';
+  import BaseCard from '../base/Card.svelte';
 
   import { copyToClipboard } from '../../lib/utils';
