+++ ohmg/frontend/svelte_components/src/components/cards/NonMapCard.svelte
@@ -1,11 +1,11 @@
 <script>
   import MapTrifold from 'phosphor-svelte/lib/MapTrifold';
 
-  import { getModal } from '../modals/BaseModal.svelte';
+  import { getModal } from '../base/Modal.svelte';
 
   import Link from '../common/Link.svelte';
 
-  import BaseCard from './BaseCard.svelte';
+  import BaseCard from '../base/Card.svelte';
 
   export let nonmap;
   export let userCanEdit;
