+++ ohmg/frontend/svelte_components/src/components/cards/SkippedCard.svelte
@@ -1,11 +1,11 @@
 <script>
   import Broom from 'phosphor-svelte/lib/Broom';
 
-  import { getModal } from '../modals/BaseModal.svelte';
+  import { getModal } from '../base/Modal.svelte';
 
   import Link from '../common/Link.svelte';
 
-  import BaseCard from './BaseCard.svelte';
+  import BaseCard from '../base/Card.svelte';
 
   export let region;
   export let sessionLocks;
