+++ ohmg/frontend/svelte_components/src/components/modals/ExtendSessionModal.svelte
@@ -1,5 +1,5 @@
 <script>
-  import Modal, { getModal } from './BaseModal.svelte';
+  import Modal, { getModal } from '../base/Modal.svelte';
   import { submitPostRequest } from '../../lib/requests';
 
   export let CONTEXT;
