+++ ohmg/frontend/svelte_components/src/components/modals/LegendModal.svelte
@@ -1,5 +1,5 @@
 <script>
-  import Modal from './BaseModal.svelte';
+  import Modal from '../base/Modal.svelte';
 
   export let id;
   export let legendUrl;
