+++ ohmg/frontend/svelte_components/src/components/modals/MapPreviewModal.svelte
@@ -1,5 +1,5 @@
 <script>
-  import Modal from './BaseModal.svelte';
+  import Modal from '../base/Modal.svelte';
   import Link from '../common/Link.svelte';
 
   export let viewerUrl = '';
